Better DNS handling for Chromium
